Changed most hooks to point to a common hook. This is because everyone just wants tab completion to  work the same:  tab does the tab according to mode  return should do newline, then tab according to mode
Attempted to make a kill buffers command.
Changed vc-root becasue it was not working agnostic
 Now it only looks for git roots.

Changed the default findcode command to be -i

Added custom compile command which saves all buffers,
 makes a compile buffer and renames it.

Added a remove dos eol chars command just because.

Tried fixing tabbing in web-mode to be 2 spaces.

Added super shift e, to eval a region but not say:
 "Region eval'd"