Added a commands derived from expand region.
collapse-list takes a list like ['a',     'b',     'c'] and collapses
any whitespace between commas.

eightyify-list takes a list and tries to make each line around 80
characters

toggle-quotes tries to toggle the quotes of any current quoted area.

collapse-list is mapped to super-a c l
eightyify-list is mapped to super-a e l
toggle-quotes is mapped to super-a t q