Update airmacs-warn to handle more cases.
Now we handle region selections better.
I removed the return check since it was annoying
I changed the variable we warn to use thing-at-point with a
sympbolic expression (sexp) instead of current-word.

I added passes to formatting to remove quites and newlines if you
warn a region.  I added a check to determine if you are inside a
string to warn with quotes.

I changed the python printing to use python3 printing
;; long live python 2 :_(